+++ R/cSpline.R
@@ -1,7 +1,7 @@
 ################################################################################
 ##
 ##   R package splines2 by Wenjie Wang and Jun Yan
-##   Copyright (C) 2016-2018
+##   Copyright (C) 2016-2019
 ##
 ##   This file is part of the R package splines2.
 ##
@@ -30,7 +30,7 @@
 ##' specified interior knots and degree, evaluated at the values of \code{x}.
 ##'
 ##' @usage
-##' cSpline(x, df = NULL, knots = NULL, degree = 3L, intercept = FALSE,
+##' cSpline(x, df = NULL, knots = NULL, degree = 3L, intercept = TRUE,
 ##'         Boundary.knots = range(x, na.rm = TRUE), scale = TRUE, ...)
 ##'
 ##' @param x The predictor variable.  Missing values are allowed and will be
@@ -46,8 +46,10 @@
 ##'     quantiles for more knots.  See also \code{Boundary.knots}.
 ##' @param degree Non-negative integer degree of the piecewise polynomial. The
 ##'     default value is 3 for cubic splines.
-##' @param intercept If \code{TRUE}, an intercept is included in the basis;
-##'     Default is \code{FALSE}.
+##' @param intercept If \code{TRUE} by default, all spline bases are included.
+##'     Notice that when using C-Spline for shape-restricted regression,
+##'     \code{intercept = TRUE} should be set even when an intercept term is
+##'     considered additional to the spline bases in the model.
 ##' @param Boundary.knots Boundary points at which to anchor the C-spline basis.
 ##'     By default, they are the range of the non-\code{NA} data.  If both
 ##'     \code{knots} and \code{Boundary.knots} are supplied, the basis
@@ -73,7 +75,7 @@
 ##' knots <- c(0.3, 0.5, 0.6)
 ##'
 ##' ### when 'scale = TRUE' (by default)
-##' csMat <- cSpline(x, knots = knots, degree = 2, intercept = TRUE)
+##' csMat <- cSpline(x, knots = knots, degree = 2)
 ##'
 ##' library(graphics)
 ##' matplot(x, csMat, type = "l", ylab = "C-spline basis")
@@ -84,12 +86,11 @@
 ##' matplot(x, msMat, type = "l", ylab = "scaled M-spline basis")
 ##'
 ##' ### when 'scale = FALSE'
-##' csMat <- cSpline(x, knots = knots, degree = 2,
-##'                  intercept = TRUE, scale = FALSE)
+##' csMat <- cSpline(x, knots = knots, degree = 2, scale = FALSE)
 ##' ## the corresponding I-splines and M-splines (with same arguments)
-##' isMat <- iSpline(x, knots = knots, degree = 2, intercept = TRUE)
+##' isMat <- iSpline(x, knots = knots, degree = 2)
 ##' msMat <- mSpline(x, knots = knots, degree = 2, intercept = TRUE)
-##' ## or using deriv methods (much more efficient)
+##' ## or using deriv methods (more efficient)
 ##' isMat1 <- deriv(csMat)
 ##' msMat1 <- deriv(csMat, derivs = 2)
 ##' ## equivalent
@@ -102,7 +103,7 @@
 ##' \code{\link{mSpline}} for M-splines.
 ##' @importFrom stats stepfun
 ##' @export
-cSpline <- function(x, df = NULL, knots = NULL, degree = 3L, intercept = FALSE,
+cSpline <- function(x, df = NULL, knots = NULL, degree = 3L, intercept = TRUE,
                     Boundary.knots = range(x, na.rm = TRUE), scale = TRUE, ...)
 {
     ## I-spline basis for inputs