+++ R/deriv.R
@@ -1,7 +1,7 @@
 ################################################################################
 ##
 ##   R package splines2 by Wenjie Wang and Jun Yan
-##   Copyright (C) 2016-2018
+##   Copyright (C) 2016-2019
 ##
 ##   This file is part of the R package splines2.
 ##
@@ -72,11 +72,11 @@
 ##' isMat <- iSpline(x, knots = knots)
 ##' stopifnot(all.equal(isMat, deriv(csMat), check.attributes = FALSE))
 ##'
-##' msMat <- mSpline(x, knots = knots)
+##' msMat <- mSpline(x, knots = knots, intercept = TRUE)
 ##' stopifnot(all.equal(msMat, deriv(isMat), deriv(csMat, 2),
 ##'                     deriv(deriv(csMat)), check.attributes = FALSE))
 ##'
-##' dmsMat <- mSpline(x, knots = knots, derivs = 1)
+##' dmsMat <- mSpline(x, knots = knots, intercept = TRUE, derivs = 1)
 ##' stopifnot(all.equal(dmsMat, deriv(msMat), deriv(isMat, 2),
 ##'                     deriv(deriv(isMat)), deriv(csMat, 3),
 ##'                     deriv(deriv(deriv(csMat))), check.attributes = FALSE))